requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
pypdf2==3.0.1
docx2txt==0.8
python-docx==1.1.0
zipfile38==0.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
